refactor: map, place 관련 파일 promise 폴더 안으로 이동
