+++ src/components/promise/SearchLocation/index.jsx
@@ -3,8 +3,8 @@ import { useEffect, useState, useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
 import Header from '../Header';
-import PlaceCardList from '@/components/place/PlaceCardList';
-import { useMapInfo } from '@/hooks/stores/map/useMapStore';
+import PlaceCardList from '@/components/promise/place/PlaceCardList';
+import { useMapInfo } from '@/hooks/stores/promise/map/useMapStore';
 import { useLocationInfo, useLocationActions } from '@/hooks/stores/promise/useLocationStore';
 import { ROUTES } from '@/constants/routes';
 import { PROMISE_LOCATION_HEADER_TEXT } from '@/constants/promise';
