+++ src/components/promise/map/MarkerManager/index.jsx
@@ -2,8 +2,8 @@ import './style.css';
 import { useEffect, useRef, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import useToggleLikePlace from '@/hooks/mutations/useToggleLikePlace';
-import { useMapInfo } from '@/hooks/stores/map/useMapStore';
-import { useMarkerInfo, useMarkerActions } from '@/hooks/stores/map/useMarkerStore';
+import { useMapInfo } from '@/hooks/stores/promise/map/useMapStore';
+import { useMarkerInfo, useMarkerActions } from '@/hooks/stores/promise/map/useMarkerStore';
 import { useUserInfo } from '@/hooks/stores/auth/useUserStore';
 import { CATEGORY, CATEGORY_MARKER_IMAGE } from '@/constants/place';
 import { MY_LOC_MARKER_IMG, MY_LOC_MARKER_ID } from '@/constants/map';
