+++ src/components/promise/map/PlaceCategoryMap/style.js
