+++ src/components/promise/place/PlaceLikeToggle/style.js
@@ -1,0 +1,18 @@
+// import * as S from './style';
+// import { theme } from '@/styles/theme';
+// import LocationIconSVG from '@/assets/img/icon/location.svg?react';
+// import EmptyHeartSvg from '@/assets/img/icon/empty_heart.svg?react';
+
+// export const Container = styled.section`
+//   display: flex;
+//   flex-direction: column;
+//   gap: 20px;
+
+//   width: 100%;
+//   padding: 10px 10px 20px;
+// `;
+
+// export const EmptyText = styled.div`
+//   margin: 30px auto;
+//   color: ${theme.color.text.grey};
+// `;
