+++ src/pages/promise/create/finalize/index.jsx
@@ -1,5 +1,5 @@
 // import * as S from './style';
-import PlaceCategoryMap from '@/components/map/PlaceCategoryMap';
+import PlaceCategoryMap from '@/components/promise/map/PlaceCategoryMap';
 
 const FinalizePage = () => {
   return <PlaceCategoryMap />;
