+++ src/stores/promise/map/markerStore.js
